+++ src/main/java/com/miaosha/demo/service/GongshuiService.java
@@ -58,6 +58,8 @@ public class GongshuiService {
     
     public static void beifen() {
     	List<Gongshui> list = disasterDao.selectAll();
+    	if(list==null) {return;}
+    	if(list.isEmpty()) return;
     	disasterDao.beifen(list);
     	disasterDao.deleteAll();
     }
