+++ src/main/java/com/miaosha/demo/service/ZhuanmuService.java
@@ -51,6 +51,8 @@ public class ZhuanmuService {
     
     public static void beifen() {
     	List<Zhuanmu> list = csdao.selectAll();
+    	if(list==null) {return;}
+    	if(list.isEmpty()) return;
     	csdao.beifen(list);
     	csdao.deleteAll();
     }
